+++ web/apps/photos/src/pages/_app.tsx
@@ -53,7 +53,7 @@ import { photosLogout } from "services/logout";
 
 // import "photoswipe/dist/photoswipe.css";
 // TODO(PS): Note, auto hide only works with the new CSS.
-import "../../../../packages/new/photos/components/ps5/dist/photoswipe.css";
+import "../../../../packages/gallery/components/viewer/ps5/dist/photoswipe.css";
 
 import "styles/global.css";
