+++ web/packages/gallery/components/viewer/.gitignore
